requirements.txt
opencv-python>=4.5.0
numpy<2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
keras-facenet>=0.3.2
tensorflow
mediapipe
protobuf<4.0.0
scipy
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt>=4.0.0
python-multipart
Pillow
email-validator